Backend/videostreaming/requirements.txt
beautifulsoup4==4.8.2
boto==2.49.0
bson==0.5.8
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangotoolbox==1.8.0
dnspython==1.16.0
docutils==0.16
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==1.1.1
get-video-properties==0.1.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
imageio==2.6.1
imageio-ffmpeg==0.4.1
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==1.2.0
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
moviepy==1.0.1
mpegdash==0.2.0
numpy==1.18.1
numpydoc==0.9.2
opencv-python==4.2.0.32
path==13.1.0
pathlib2==2.3.5
Pillow==7.0.0
pymongo==3.10.1
PyMySQL==0.9.3
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-ffmpeg-video-streaming==0.0.25
python-googlegeocoder==0.4.0
python-mpd==0.3.0
PyYAML==5.3
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.2.4
unicodecsv==0.14.1
urllib3==1.25.8
virtualenv==20.0.8
virtualenvwrapper-win==1.2.6